requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pymysql==1.1.1
python-dotenv==1.0.1
pandas>=1.5.0
pymongo>=4.0.0
requests>=2.0.0
python-dotenv>=0.21.0
pydantic==2.9.2
pymongo==4.6.1

